+++ go/stream_commands_test.go
@@ -492,7 +492,7 @@ func ExampleClient_XReadGroupWithOptions() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-1
-	// Entry fields: map[entry1_field1:entry1_value1 entry1_field2:entry1_value2]
+	// Entry fields: [{entry1_field1 entry1_value1} {entry1_field2 entry1_value2}]
 }
 
 func ExampleClusterClient_XReadGroupWithOptions() {
@@ -534,7 +534,7 @@ func ExampleClusterClient_XReadGroupWithOptions() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-1
-	// Entry fields: map[entry1_field1:entry1_value1 entry1_field2:entry1_value2]
+	// Entry fields: [{entry1_field1 entry1_value1} {entry1_field2 entry1_value2}]
 }
 
 func ExampleClient_XRead() {
@@ -571,7 +571,7 @@ func ExampleClient_XRead() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-1
-	// Entry fields: map[field1:value1 field2:value2]
+	// Entry fields: [{field1 value1} {field2 value2}]
 }
 
 func ExampleClusterClient_XRead() {
@@ -608,7 +608,7 @@ func ExampleClusterClient_XRead() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-1
-	// Entry fields: map[field1:value1 field2:value2]
+	// Entry fields: [{field1 value1} {field2 value2}]
 }
 
 func ExampleClient_XReadWithOptions() {
@@ -655,7 +655,7 @@ func ExampleClient_XReadWithOptions() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-2
-	// Entry fields: map[field3:value3 field4:value4]
+	// Entry fields: [{field3 value3} {field4 value4}]
 }
 
 func ExampleClusterClient_XReadWithOptions() {
@@ -702,7 +702,7 @@ func ExampleClusterClient_XReadWithOptions() {
 	// Stream exists: true
 	// Number of entries: 1
 	// Entry ID: 12345-2
-	// Entry fields: map[field3:value3 field4:value4]
+	// Entry fields: [{field3 value3} {field4 value4}]
 }
 
 func ExampleClient_XDel() {
@@ -1477,8 +1477,8 @@ func ExampleClient_XClaimWithOptions() {
 		fmt.Printf("Message ID: %s with retry count: %d\n", id, 3)
 		// Print a sample field if available
 		if len(claimResponse.Fields) > 0 {
-			for field, value := range claimResponse.Fields {
-				fmt.Printf("Field: %s, Value: %s\n", field, value)
+			for _, keyValue := range claimResponse.Fields {
+				fmt.Printf("Field: %s, Value: %s\n", keyValue.Key, keyValue.Value)
 				break // Just print the first field as an example
 			}
 		}
@@ -1936,10 +1936,8 @@ func ExampleClient_XInfoStream() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	jsonResult, _ := json.MarshalIndent(response, "", "  ")
 
-	fmt.Println(string(jsonResult))
-	// Output:
+	// Response Structure is as follows:
 	// {
 	//   "entries-added": 1,
 	//   "first-entry": [
@@ -1964,6 +1962,18 @@ func ExampleClient_XInfoStream() {
 	//   "radix-tree-nodes": 2,
 	//   "recorded-first-entry-id": "12345-1"
 	// }
+
+	// Output a few entries from the return object.
+	fmt.Printf("Entries Added: %d\n", response.EntriesAdded)
+	fmt.Printf("Groups: %d\n", response.Groups)
+	fmt.Printf("Last generated Id: %s\n", response.LastGeneratedID)
+	fmt.Printf("Length: %d\n", response.Length)
+
+	// Output:
+	// Entries Added: 1
+	// Groups: 0
+	// Last generated Id: 12345-1
+	// Length: 1
 }
 
 func ExampleClusterClient_XInfoStream() {
@@ -1981,10 +1991,8 @@ func ExampleClusterClient_XInfoStream() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	jsonResult, _ := json.MarshalIndent(response, "", "  ")
 
-	fmt.Println(string(jsonResult))
-	// Output:
+	// Response Structure is as follows:
 	// {
 	//   "entries-added": 1,
 	//   "first-entry": [
@@ -2009,6 +2017,18 @@ func ExampleClusterClient_XInfoStream() {
 	//   "radix-tree-nodes": 2,
 	//   "recorded-first-entry-id": "12345-1"
 	// }
+
+	// Output a few entries from the return object.
+	fmt.Printf("Entries Added: %d\n", response.EntriesAdded)
+	fmt.Printf("Groups: %d\n", response.Groups)
+	fmt.Printf("Last generated Id: %s\n", response.LastGeneratedID)
+	fmt.Printf("Length: %d\n", response.Length)
+
+	// Output:
+	// Entries Added: 1
+	// Groups: 0
+	// Last generated Id: 12345-1
+	// Length: 1
 }
 
 func ExampleClient_XInfoStreamFullWithOptions() {